% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ComicPanelR-package.R
\docType{package}
\name{ComicPanelR-package}
\alias{ComicPanelR}
\alias{ComicPanelR-package}
\title{ComicPanelR: Create Comic Page Layouts with Panels}
\description{
The packacge allows to specify the layout of comic pages with panels. It converts layout definitions to image files. In particular, stencils for individual panels can be created.
}
\author{
\strong{Maintainer}: Christof Schötz \email{math@christof-schoetz.de} (\href{https://orcid.org/0000-0003-3528-4544}{ORCID})

}
\keyword{internal}
