% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coor2image.R
\name{coor2image}
\alias{coor2image}
\title{Convert Coordinates File to Image}
\usage{
coor2image(fileName)
}
\arguments{
\item{fileName}{The base file name (preferably without file extension).
".coor.json" will be added to obtain the input file. The json file should
contain an object of class \verb{ ["Polygon", "List", "Opts"]}.}
}
\description{
Convert Coordinates File to Image
}
