% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stencil.R
\name{coor2panelStencils}
\alias{coor2panelStencils}
\title{Convert Coordinates One Stencil Image per Panel}
\usage{
coor2panelStencils(fileName)
}
\arguments{
\item{fileName}{The base file name (preferably without file extension).
".coor.json" will be added to obtain the input file. The json file should
contain an object of class \verb{ ["Polygon", "List", "Opts"]}.}
}
\description{
Convert Coordinates One Stencil Image per Panel
}
