% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout2coor.R
\name{layout2coor}
\alias{layout2coor}
\title{Convert a Layout Specification File into a Coordinates File}
\usage{
layout2coor(fileName)
}
\arguments{
\item{fileName}{The base file name (preferably without file extension).
".layout.json" will be added to obtain the input file. The json file should
contain an object of class \verb{ ["Comic", "Opts"]}.}
}
\description{
Convert a Layout Specification File into a Coordinates File
}
